requirements.txt
fastapi~=0.73.0
pymongo[srv]~=4.0.1
python-dotenv~=0.19.2
uvicorn[standard]~=0.17.1
